requirements.txt
adagio==0.2.4
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.1.0
audioread==3.0.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comtypes==1.4.5
contourpy==1.1.1
convertdate==2.4.0
coreforecast==0.0.12
cycler==0.12.1
Cython==3.0.10
darts==0.30.0
decorator==5.1.1
ephem==4.1.5
filelock==3.15.4
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.1
google-auth==2.29.0
holidays==0.53
html5lib==1.1
huggingface-hub==0.0.8
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.6
llvmlite==0.41.1
LunarCalendar==0.0.9
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
networkx==3.1
nfoursid==1.0.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
packaging==24.1
pandas==1.5.3
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pmdarima==2.0.4
pooch==1.8.2
protobuf==5.27.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
PyMeeus==0.5.12
pyod==2.0.1
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.3.3
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.3
scikit-base==0.7.8
scikit-learn==1.3.2
scipy==1.10.1
shap==0.44.1
six==1.16.0
sktime==0.29.1
slicer==0.0.7
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.4
statsforecast==1.7.6
statsmodels==0.14.1
sympy==1.13.1
tbats==1.1.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==1.4.1
tqdm==4.66.4
transformers==4.6.1
triad==0.9.8
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
urllib3==2.2.1
utilsforecast==0.2.2
webencodings==0.5.1
Werkzeug==3.0.2
xarray==2023.1.0
xgboost==2.1.1
yarl==1.9.4
yfinance==0.2.41
zipp==3.18.1
